refactor component to include selecting chart options before chart
